add possibility to change app theme add fileslist view mode rename some settings stuff use enums instead of strings more performance improvements other things i dont remember

+++ lib/api_definitions.dart
@@ -100,7 +100,7 @@ class UploadingEvent {}
 
 class UploadingEventProgress extends UploadingEvent {
   double progress;
-  double bytesPerSec;
+  int bytesPerSec;
 
   UploadingEventProgress({
     required this.progress,
